+++ lib/glimmer/libui/control_proxy.rb
@@ -167,18 +167,18 @@ module Glimmer
       end
       
       def handle_listener(listener_name, &listener)
-        safe_listener = Proc.new { listener.call(self) }
+        safe_listener = Proc.new { |*args| listener.call(self, *args[1..-1]) }
         if ::LibUI.respond_to?("#{libui_api_keyword}_#{listener_name}")
           if listeners[listener_name].nil?
-            ::LibUI.send("#{libui_api_keyword}_#{listener_name}", @libui) do
-              listeners_for(listener_name).map { |listener| listener.call }.last
+            ::LibUI.send("#{libui_api_keyword}_#{listener_name}", libui) do |*args|
+              listeners_for(listener_name).map { |listener| listener.call(*args) }.last
             end
           end
           listeners_for(listener_name) << safe_listener
         elsif ::LibUI.respond_to?("control_#{listener_name}")
           if listeners[listener_name].nil?
-            ::LibUI.send("control_#{listener_name}", @libui) do
-              listeners_for(listener_name).map { |listener| listener.call }.last
+            ::LibUI.send("control_#{listener_name}", libui) do |*args|
+              listeners_for(listener_name).map { |listener| listener.call(*args) }.last
             end
           end
           listeners_for(listener_name) << safe_listener
@@ -272,32 +272,36 @@ module Glimmer
       def send_to_libui(method_name, *args, &block)
         if ::LibUI.respond_to?("#{libui_api_keyword}_#{method_name.to_s.sub(/\?$/, '')}") && args.empty?
           property = method_name.to_s.sub(/\?$/, '')
-          value = ::LibUI.send("#{libui_api_keyword}_#{property}", @libui, *args)
+          value = ::LibUI.send("#{libui_api_keyword}_#{property}", libui, *args)
+          handle_string_property(property, handle_boolean_property(property, value))
+        elsif ::LibUI.respond_to?("#{libui_api_keyword}_get_#{method_name.to_s.sub(/\?$/, '')}") && args.empty?
+          property = method_name.to_s.sub(/\?$/, '')
+          value = ::LibUI.send("#{libui_api_keyword}_get_#{property}", libui, *args)
           handle_string_property(property, handle_boolean_property(property, value))
         elsif ::LibUI.respond_to?("#{libui_api_keyword}_set_#{method_name.to_s.sub(/=$/, '')}") && !args.empty?
           property = method_name.to_s.sub(/=$/, '')
           args[0] = Glimmer::LibUI.boolean_to_integer(args.first) if BOOLEAN_PROPERTIES.include?(property) && (args.first.is_a?(TrueClass) || args.first.is_a?(FalseClass))
           args[0] = '' if STRING_PROPERTIES.include?(property) && args.first == nil
           if property.to_s == 'checked'
-            current_value = Glimmer::LibUI.integer_to_boolean(::LibUI.send("#{libui_api_keyword}_checked", @libui), allow_nil: false)
+            current_value = Glimmer::LibUI.integer_to_boolean(::LibUI.send("#{libui_api_keyword}_checked", libui), allow_nil: false)
             new_value = Glimmer::LibUI.integer_to_boolean(args[0], allow_nil: false)
-            ::LibUI.send("#{libui_api_keyword}_set_#{property}", @libui, *args) if new_value != current_value
+            ::LibUI.send("#{libui_api_keyword}_set_#{property}", libui, *args) if new_value != current_value
           else
-            ::LibUI.send("#{libui_api_keyword}_set_#{property}", @libui, *args)
+            ::LibUI.send("#{libui_api_keyword}_set_#{property}", libui, *args)
           end
         elsif ::LibUI.respond_to?("#{libui_api_keyword}_#{method_name}") && !args.empty?
-          ::LibUI.send("#{libui_api_keyword}_#{method_name}", @libui, *args)
+          ::LibUI.send("#{libui_api_keyword}_#{method_name}", libui, *args)
         elsif ::LibUI.respond_to?("control_#{method_name.to_s.sub(/\?$/, '')}") && args.empty?
           property = method_name.to_s.sub(/\?$/, '')
-          value = ::LibUI.send("control_#{property}", @libui, *args)
+          value = ::LibUI.send("control_#{property}", libui, *args)
           handle_string_property(property, handle_boolean_property(property, value))
         elsif ::LibUI.respond_to?("control_set_#{method_name.to_s.sub(/=$/, '')}")
           property = method_name.to_s.sub(/=$/, '')
           args[0] = Glimmer::LibUI.boolean_to_integer(args.first) if BOOLEAN_PROPERTIES.include?(property) && (args.first.is_a?(TrueClass) || args.first.is_a?(FalseClass))
           args[0] = '' if STRING_PROPERTIES.include?(property) && args.first == nil
-          ::LibUI.send("control_set_#{method_name.to_s.sub(/=$/, '')}", @libui, *args)
+          ::LibUI.send("control_set_#{method_name.to_s.sub(/=$/, '')}", libui, *args)
         elsif ::LibUI.respond_to?("control_#{method_name}") && !args.empty?
-          ::LibUI.send("control_#{method_name}", @libui, *args)
+          ::LibUI.send("control_#{method_name}", libui, *args)
         end
       end